homework 8

+++ controllers/TaskController.php
@@ -6,7 +6,9 @@ use app\models\query\TaskQuery;
 use Yii;
 use app\models\Task;
 use app\models\search\TaskSearch;
+use yii\data\ActiveDataProvider;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\filters\AccessControl;
@@ -59,6 +61,42 @@ class TaskController extends Controller
     }
 
     /**
+     * Lists all Task models.
+     * @return mixed
+     */
+    public function actionShared()
+    {
+        $searchModel = new TaskSearch();
+        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        /** @var $query TaskQuery */
+        $query = $dataProvider->query;
+        $query->byCreator(Yii::$app->user->id)
+            ->innerJoinWith(Task::RELATION_TASK_USERS);
+        return $this->render('shared', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+        ]);
+    }
+
+    /**
+     * Lists all Task models.
+     * @return mixed
+     */
+    public function actionAccessed()
+    {
+        $searchModel = new TaskSearch();
+        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        /** @var $query TaskQuery */
+        $query = $dataProvider->query;
+        $query->innerJoinWith(Task::RELATION_TASK_USERS)
+            ->where(['user_id' => Yii::$app->user->id]);
+        return $this->render('accessed', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+        ]);
+    }
+
+    /**
      * Displays a single Task model.
      * @param integer $id
      * @return mixed
@@ -66,8 +104,14 @@ class TaskController extends Controller
      */
     public function actionView($id)
     {
+        $model = $this->findModel($id);
+        $dataProvider = new ActiveDataProvider([
+            'query' => $model->getTaskUsers(),
+        ]);
+
         return $this->render('view', [
             'model' => $this->findModel($id),
+            'dataProvider' => $dataProvider,
         ]);
     }
 
@@ -81,7 +125,8 @@ class TaskController extends Controller
         $model = new Task();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            Yii::$app->session->setFlash('success', 'Задача создана.');
+            return $this->redirect(['task/my']);
         }
 
         return $this->render('create', [
@@ -95,13 +140,21 @@ class TaskController extends Controller
      * @param integer $id
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
+     * @throws ForbiddenHttpException
      */
     public function actionUpdate($id)
     {
+        $task = Task::findOne($id);
+
+        if($task->creator_id != Yii::$app->user->id) {
+            throw new ForbiddenHttpException('Нет доступа.');
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            Yii::$app->session->setFlash('success', 'Задача обновлена.');
+            return $this->redirect(['task/my']);
         }
 
         return $this->render('update', [
@@ -115,12 +168,21 @@ class TaskController extends Controller
      * @param integer $id
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
+     * @throws ForbiddenHttpException
      */
     public function actionDelete($id)
     {
+        $task = Task::findOne($id);
+
+        if($task->creator_id != Yii::$app->user->id) {
+            throw new ForbiddenHttpException('Нет доступа.');
+        }
+
         $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+        Yii::$app->session->setFlash('success', 'Задача удалёна.');
+
+        return $this->redirect(['task/my']);
     }
 
     /**
